requirements.txt
beautifulsoup4
emoji
fastapi
google-api-python-client
nltk
numpy
openai
pandas
pydantic
python-dotenv
requests
scikit-learn
spacy
textblob
torch
transformers
watchdog
youtube-transcript-api
